- patch from Florian Falkner, with some modifications (breaks Multiline_Editbox) - uncritical compile time warning in computer player fixed
git-svn-id: https://widelands.svn.sourceforge.net/svnroot/widelands/trunk@837 37b2a8de-5219-0410-9f54-a31bc463ab9c

+++ src/computer_player.cc
@@ -630,7 +630,7 @@ bool Computer_Player::improve_roads (Flag* flag)
 
 	CheckStepRoadAI check(player, MOVECAPS_WALK, false);
 	
-	for (i=1;i<nearflags.size();i++) {
+	for (i=1;i< (int)nearflags.size();i++) {
 	    NearFlag& nf=nearflags[i];
 	    
 	    if (2*nf.distance+2>=nf.cost)
